+++ src/main/java/com/example/springbootmonolith/repositories/SongRepository.java
@@ -1,7 +1,9 @@
 package com.example.springbootmonolith.repositories;
 
 import com.example.springbootmonolith.models.Song;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SongRepository extends CrudRepository<Song, Integer> {
+public interface SongRepository extends CrudRepository<Song, Long> {
+
 }
